create generic method to decode data from network request
